requirements.txt
pymemcache
attrs